// src/tasks/jobs.types.ts
export type TasksPayload = {
  jobs: TasksJobs;
  crons: { task: string; match: string }[];
};
export type TasksJobs = Record<string, TaskJob>;

type TaskFn = (payload: Record<string, any>) => Promise<void>;

export type TaskJob = {
  action: TaskFn;
  longRunning?: boolean;
};